Added Adjustment factor
The Earlier version involved a bug such that if the end date was passed as 31-3-2020, The prices returned will be till 29-3-2020. The last day was excluded. The -1 on line 115 was causing the above issue. The -1 earlier was to naturally include the start date as the yahoo finance database ignore PERIOD1 by design.

Moreover, it seems that yahoo finance no longer requires the cookies to validate a session before it provides the data. So the cookies and crumb is empty in version 2.0