New Feature in version 2.0
DynamicCalculation is an interface whose object can be passed to ServerConnect. Server connect reads each line from the input stream and passes the line in string format to the method calculate. The idea is that the user of the API can define calculate and be able to make computations at the time the data is being fetched. This allows version 2.0 to not save the files later (see description of ServerConnect commit to know how that is done). It also reduces the time complexity of the entire application by not having to open the file separately, iterate through it and make the computations.