Added New Feature in 2.0
The ServerConnect class now accepts DynamicCalculation object in its constructor to enable the user of the API to define a custom calculation to be made on each row of equity price data. It is non-mandatory for the user of the API to save the data being extracted. The user can either implement the DynamicCalculation or save the data in a .csv file. In order to not save the file, pass a null target object to getCSV().